requirements.txt
requests
docx2pdf
reportlab
python-docx
genai
pandoc